renaming and tidying
