+++ Neo4j.Driver/Neo4j.Driver.Tests.BenchkitBackend/Abstractions/IWorkloadExecutorSelector.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) "Neo4j"
 // Neo4j Sweden AB [https://neo4j.com]
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License").
 // You may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,7 +17,14 @@ using Neo4j.Driver.Tests.BenchkitBackend.Types;
 
 namespace Neo4j.Driver.Tests.BenchkitBackend.Abstractions;
 
-internal interface IWorkloadExecutor
+/// <summary>
+/// Methods for executing a workload.
+/// </summary>
+public interface IWorkloadExecutorSelector
 {
-    Task ExecuteWorkloadAsync(Workload workload);
+    /// <summary>
+    /// Execute a workload.
+    /// </summary>
+    /// <param name="workload">The workload to execute.</param>
+    public Task ExecuteWorkloadAsync(Workload workload);
 }
