+++ Neo4j.Driver/Neo4j.Driver.Tests.BenchkitBackend/Abstractions/IWorkloadStore.cs
@@ -1,23 +1,26 @@
 ﻿// Copyright (c) "Neo4j"
 // Neo4j Sweden AB [https://neo4j.com]
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License").
 // You may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Neo4j.Driver.Tests.BenchkitBackend.Implementations;
+using Neo4j.Driver.Tests.BenchkitBackend.Types;
 
 namespace Neo4j.Driver.Tests.BenchkitBackend.Abstractions;
 
-internal interface IWorkloadExecutionMethod
+internal interface IWorkloadStore
 {
-    Task ExecuteWorkloadAsync(Workload workload);
+    string  CreateWorkload(Workload workload);
+    Workload GetWorkload(string id);
+    Workload UpdateWorkload(string id, Workload workload);
+    void DeleteWorkload(string id);
 }
