+++ Neo4j.Driver/Neo4j.Driver.Tests.BenchkitBackend/Types/Workload.cs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Neo4j.Driver.Tests.BenchkitBackend.Implementations;
+namespace Neo4j.Driver.Tests.BenchkitBackend.Types;
 
 public enum Method
 {
